Replace usage of 'as' keyword
It hides that the conversion may be lossy.

This requires breaking ABI for the FFI interface, because the old
behaviour would truncate the number of masters if a plugin had more than
255 (which is not valid in-game, but esplugin could be given an invalid
plugin). This would ultimately cause memory management issues as the
caller wouldn't be able to pass back the full size of the array when it
came to freeing it.

The remaining usages were all to do with comparing with usize values
or indexing, and have been replaced by a function that will panic if the
u32 -> usize conversion will fail but which optimises away on platforms
where usize >= u32 (which is everywhere I expect this code to
run).